+++ lib/many_lints.dart
@@ -35,6 +35,7 @@ import 'package:many_lints/src/rules/avoid_passing_bloc_to_bloc.dart';
 import 'package:many_lints/src/rules/avoid_passing_build_context_to_blocs.dart';
 import 'package:many_lints/src/rules/prefer_bloc_extensions.dart';
 import 'package:many_lints/src/rules/prefer_immutable_bloc_state.dart';
+import 'package:many_lints/src/rules/prefer_immutable_state.dart';
 import 'package:many_lints/src/rules/prefer_multi_bloc_provider.dart';
 import 'package:many_lints/src/rules/avoid_cascade_after_if_null.dart';
 import 'package:many_lints/src/rules/avoid_conditional_hooks.dart';
@@ -244,7 +245,7 @@ import 'package:many_lints/src/fixes/avoid_single_field_destructuring_fix.dart';
 import 'package:many_lints/src/fixes/avoid_commented_out_code_fix.dart';
 import 'package:many_lints/src/fixes/avoid_incomplete_copy_with_fix.dart';
 import 'package:many_lints/src/fixes/prefer_bloc_extensions_fix.dart';
-import 'package:many_lints/src/fixes/prefer_immutable_bloc_state_fix.dart';
+import 'package:many_lints/src/fixes/add_immutable_annotation_fix.dart';
 import 'package:many_lints/src/fixes/prefer_multi_bloc_provider_fix.dart';
 import 'package:many_lints/src/fixes/avoid_incorrect_image_opacity_fix.dart';
 import 'package:many_lints/src/fixes/avoid_unnecessary_gesture_detector_fix.dart';
@@ -352,6 +353,7 @@ class ManyLintsPlugin extends Plugin {
     _registerWarningRule(registry, AvoidPassingBuildContextToBlocs());
     _registerWarningRule(registry, PreferBlocExtensions());
     _registerWarningRule(registry, PreferImmutableBlocState());
+    _registerWarningRule(registry, PreferImmutableState());
     _registerWarningRule(registry, PreferMultiBlocProvider());
     _registerWarningRule(registry, AvoidCascadeAfterIfNull());
     _registerWarningRule(registry, AvoidCommentedOutCode());
@@ -891,7 +893,11 @@ class ManyLintsPlugin extends Plugin {
     );
     registry.registerFixForRule(
       PreferImmutableBlocState.code,
-      PreferImmutableBlocStateFix.new,
+      AddImmutableAnnotationFix.new,
+    );
+    registry.registerFixForRule(
+      PreferImmutableState.code,
+      AddImmutableAnnotationFix.new,
     );
     registry.registerFixForRule(
       PreferMultiBlocProvider.code,